ADD NAVBAR, SIDEBAR AND ROUTING
This is still the barebones of the project frontend.

The interface is made mainly with Angular Material.

The navbar and sidebar follow  quite closely the prototypes made.
Though they are not functional in most cases.
Backend development is necessary.

The routing was added and seems to work well.
Yet the pages that are routed are just tests for what's upcoming.

Yarn is been used as the package manager instead of npm.
Due to it's faster install speed for multiple packages.

SCSS is been used for custom theming.
Aside the advantages of it over normal CSS.